更新训练脚本以支持动态检查点路径，调整数据集配置以使用新的BEDLAM子集，优化重投影损失计算，添加可视化功能以提高训练过程的可解释性。

+++ mega/train/train_conditional_vqmae/train.py
@@ -231,7 +231,7 @@ class CVQMAE_Train(Train):
             reproj_loss = 0
             if is_3dpw.any(): # 3DPW/EMDB/BEDLAM
                 reproj_loss += reprojection_loss(
-                    data["j2d"][is_3dpw][:, :, :2], # [:, :, :2]
+                    data["j2d"][is_3dpw][:, :, :2].to(torch.float32), # [:, :, :2]
                     pred_mesh[is_3dpw],
                     pred_cam[is_3dpw],
                     self.joints_reg_smpl,
@@ -512,7 +512,7 @@ class CVQMAE_Train(Train):
             reproj_loss = 0
             if is_3dpw.any():
                 reproj_loss += reprojection_loss(
-                    data["j2d"][is_3dpw][:, :, :2],
+                    data["j2d"][is_3dpw][:, :, :2].to(torch.float32),
                     pred_mesh[is_3dpw],
                     pred_cam[is_3dpw],
                     self.joints_reg_smpl,
